remove warning
git-svn-id: https://svn.cern.ch/reps/SPIDR/software/trunk@4947 4525493e-7705-40b1-a816-d608a930855b

+++ SpidrTpx3Lib/SpidrDaq.cpp
@@ -200,7 +200,7 @@ bool SpidrDaq::startRecording( std::string filename,
   memset( static_cast<void *>(&fhdr->descr), 0, sizeof(fhdr->descr) );
   if( !descr.empty() )
     {
-      u32 sz = descr.length();
+      size_t sz = descr.length();
       if( sz > sizeof(fhdr->descr)-1 )
 	sz = sizeof(fhdr->descr)-1;
